Extract bind group layout structs to their own file
* rename `BindingDescriptor` to `BindGroupEntry`
