+++ src/renderer/src/shader/shader_reflect.rs
@@ -1,6 +1,6 @@
 use spirv_reflect::{ShaderModule, types::ReflectDescriptorBinding, types::ReflectDescriptorSet, types::{ReflectDescriptorType, ReflectInterfaceVariable}};
 
-use crate::{BindGroupDescriptor, BindingDescriptor, BindingType, Shader, ShaderStage, UniformProperty, VertexAttributeDescriptor, WgpuInto};
+use crate::{BindGroupDescriptor, BindGroupEntry, BindingType, Shader, ShaderStage, UniformProperty, VertexAttributeDescriptor, WgpuInto};
 
 /// A ShaderLayout describes the layout of the loaded shader, analyzed by reflection.
 ///
@@ -24,6 +24,7 @@ pub(crate) fn reflect(spv_data: &[u8]) -> ShaderLayout {
     let shader_stage: ShaderStage = shader_module.get_shader_stage().into();
 
     let bind_groups: Vec<BindGroupDescriptor> = reflect_bind_groups(&shader_module, shader_stage);
+    dbg!(&bind_groups);
     let _input_variables: Vec<VertexAttributeDescriptor> = reflect_input_variables(&shader_module);
 
     reflect_push_constant_blocks(&shader_module);
@@ -46,7 +47,7 @@ pub(crate) fn reflect_bind_group(
     descriptor_set: &ReflectDescriptorSet,
     shader_stage: ShaderStage,
 ) -> BindGroupDescriptor {
-    let bindings: Vec<BindingDescriptor> = descriptor_set.bindings
+    let bindings: Vec<BindGroupEntry> = descriptor_set.bindings
         .iter()
         .map(|binding| {
             reflect_binding(binding, shader_stage)
@@ -58,8 +59,8 @@ pub(crate) fn reflect_bind_group(
 pub(crate) fn reflect_binding(
     binding: &ReflectDescriptorBinding,
     shader_stage: ShaderStage,
-) -> BindingDescriptor {
-    println!("BINDING: {:?}", binding);
+) -> BindGroupEntry {
+    // println!("BINDING: {:?}", binding);
 
     let type_description = binding.type_description.as_ref().unwrap();
     let (name, binding_type) = match binding.descriptor_type {
@@ -73,7 +74,7 @@ pub(crate) fn reflect_binding(
         _ => panic!("Unsupported binding type {:?}", binding.descriptor_type),
     };
 
-    BindingDescriptor {
+    BindGroupEntry {
         name: name.to_string(),
         index: binding.binding,
         binding_type,
